Minor improvement

+++ app/db/engine.py
@@ -18,7 +18,7 @@ def create_database_or_engine(db_url: str, db_name: str, options: dict
         _create_database(db_url, db_name)
         log.warning(f'Creating database {db_name}')
     except OperationalError:
-        log.warning(f'Database {db_name} already EXIST')
+        pass
 
     return create_engine(f'{db_url}/{db_name}', **options)
 
@@ -35,7 +35,10 @@ engine = create_database_or_engine(BASE_URL, DB_NAME, ENGINE_OPTIONS)
 def drop_database(db_path: str, db_name: str) -> None:
     """This function delete database by db_url and db_name"""
     my_db = f'{db_path}/{db_name}'
-    os.remove(my_db)
+    if os.path.exists(my_db):
+        os.remove(my_db)
+    else:
+        log.warning(f"Database {db_name} don't exist")
 
 
 def create_table(db_engine: Engine = engine) -> None:
